FIX: type conversion warnings

+++ src/applications/mpt-daemons/receiver/receiverd_worker.c
@@ -150,7 +150,7 @@ static void *receiverd_handle_stats(const maestro_note_t *request,
 
 	maestro_note_stats_set_perf_count(response, total_msg);
 	maestro_note_stats_set_perf_rate(response, (total_rate / count_children));
-	maestro_note_stats_set_perf_latency(response, (total_lat / count_children));
+	maestro_note_stats_set_perf_latency(response, ((double) total_lat / count_children));
 
 	return NULL;
 }
